[fix]: [CCM-5379]: metrics: restore default labels assigned by opencensus lib. (#29879)
Open-census library, by default, adds a label named `opencensus_task`
to each of the collected measurement. The value of the label is
pod name in our case.

This behaviour was causing an issue with `LastValue` type of
metrics because each distinct value of the `opencensus_task` would create
a separate time-series and hence a separate `LastValue` in each time-series.

In the 9c75c43a402728fbb5472b19cdd2057d9bf4995a commit, I had made a
one liner change to disable this behaviour.

The change has a side-effect: each unique combination of
label values (e.g. <accountId, clusterId, messageType>) identifies a unique time-series.
Without the task label, it is now possible that two different pods try to publish
to the same time-series leading to exception.

This patch restores the earlier behaviour of the opencensus library
to publish default set of labels.