fix:[CDS-52541]: Silence the mapping exception behind FF. (#43521)
We are unable to RCA a corner case during map operation. The method throw the error "Converter org.modelmapper.internal.converter.CollectionConverter@354aaffe failed to convert com.mongodb.BasicDBList to java.util.List". We choose to iterate the source and try to map each element to the same target, using this strategy when the issue comes again we will have more details.