fix:[CDS-69502]: Add missing exclamation mark (#48130)
To evaluate whether we should use cache or not, evaluating the helm version to V3 and FF should be not enabled. Other places from the code have the same condition, and here the exclamation mark was missing.

Without this, after the execution, the cleanup rules are not executed and left garbage on the disk.