fix:[CDS-71319]: Add exception when one is missing inside the error response data (#52271)
* fix:[CDS-71319]: Add exception to error response data when task fails during validation phase

An NPE is regularly thrown when a delegate task fails during the validation phase. In general, a validation failure does not need to be related to an exception, but from a notify event listener a failure always needs to be related to one.

* fix:[CDS-71319]: Add exception when one is missing inside the error response data

An NPE is regularly thrown inside the GitCommandCallback when an exception is missing over the NotifyEventListener process. From a notify event listener perspective a failure always needs to be related to one exception.

* fix:[CDS-71319]: Add entry to error code message property file